refactor: update payment module
